+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 ## Versions
 
+- [V5.6.1](releases.d/RELEASE_NOTES_V5.6.1.md)
 - [V5.6.0](releases.d/RELEASE_NOTES_V5.6.0.md)
 - [V5.5.1](releases.d/RELEASE_NOTES_V5.5.1.md)
 - [V5.5.0](releases.d/RELEASE_NOTES_V5.5.0.md)